notebooks/hf-dog-breed-demo/requirement.txt
tensorflow==2.15.0
gradio
numpy
pillow
